# src/artifacts/swagger.json
{"openapi":"3.0.0","paths":{"/api/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Usuário criado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Users"}}}},"400":{"description":"Email já está cadastrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestException"}}}}},"summary":"Cria um novo usuário","tags":["users"]},"get":{"operationId":"UsersController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Users"}}}}}},"summary":"Pega todos os usuários existentes","tags":["users"]}},"/api/users/{id}":{"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Usuário atualizado com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Users"}}}},"404":{"description":"Usuário não encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}}},"summary":"Atualiza um usuário pelo seu ID","tags":["users"]}}},"info":{"title":"Users CRUD","description":"API de usuários","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}},"required":["firstName","lastName","email"]},"Users":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}},"required":["firstName","lastName","email"]},"BadRequestException":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["firstName","lastName"]},"NotFoundException":{"type":"object","properties":{}}}}}